+++ rbx_binary/src/deserializer.rs
@@ -1,5 +1,5 @@
 use std::{
-    collections::{HashMap, VecDeque},
+    collections::{HashMap, HashSet, VecDeque},
     convert::TryInto,
     io::{self, Read},
     str,
@@ -18,6 +18,7 @@ use rbx_reflection::DataType;
 use thiserror::Error;
 
 use crate::{
+    cframe,
     chunk::Chunk,
     core::{
         find_canonical_property_descriptor, RbxReadExt, FILE_MAGIC_HEADER, FILE_SIGNATURE,
@@ -88,139 +89,19 @@ pub(crate) enum InnerError {
     #[error("File referred to type ID {type_id}, which was not declared")]
     InvalidTypeId { type_id: u32 },
 
-    #[error("Invalid property data: CFrame property {type_name}.{prop_name} had an invalid orientation ID {id:02x}")]
-    BadCFrameOrientationId {
+    #[error("Invalid property data: CFrame property {type_name}.{prop_name} had an invalid rotation ID {id:02x}")]
+    BadRotationId {
         type_name: String,
         prop_name: String,
         id: u8,
     },
-}
 
-// TODO potentially move this to a different file if/when we do the inverse for serializing
-pub(crate) fn special_case_to_rotation(id: u8) -> Option<Matrix3> {
-    match id {
-        0x02 => Some(Matrix3::new(
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-        )),
-        0x03 => Some(Matrix3::new(
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(0.0, 1.0, 0.0),
-        )),
-        0x05 => Some(Matrix3::new(
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-        )),
-        0x06 => Some(Matrix3::new(
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(0.0, -1.0, 0.0),
-        )),
-        0x07 => Some(Matrix3::new(
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-        )),
-        0x09 => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, 1.0, 0.0),
-        )),
-        0x0a => Some(Matrix3::new(
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-        )),
-        0x0c => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, -1.0, 0.0),
-        )),
-        0x0d => Some(Matrix3::new(
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(1.0, 0.0, 0.0),
-        )),
-        0x0e => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(1.0, 0.0, 0.0),
-        )),
-        0x10 => Some(Matrix3::new(
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(1.0, 0.0, 0.0),
-        )),
-        0x11 => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(1.0, 0.0, 0.0),
-        )),
-        0x14 => Some(Matrix3::new(
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-        )),
-        0x15 => Some(Matrix3::new(
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(0.0, 1.0, 0.0),
-        )),
-        0x17 => Some(Matrix3::new(
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-        )),
-        0x18 => Some(Matrix3::new(
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(0.0, -1.0, 0.0),
-        )),
-        0x19 => Some(Matrix3::new(
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-        )),
-        0x1b => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, 1.0, 0.0),
-        )),
-        0x1c => Some(Matrix3::new(
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-        )),
-        0x1e => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-            Vector3::new(0.0, -1.0, 0.0),
-        )),
-        0x1f => Some(Matrix3::new(
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-        )),
-        0x20 => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-        )),
-        0x22 => Some(Matrix3::new(
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-        )),
-        0x23 => Some(Matrix3::new(
-            Vector3::new(0.0, 0.0, -1.0),
-            Vector3::new(0.0, -1.0, 0.0),
-            Vector3::new(-1.0, 0.0, 0.0),
-        )),
-        _ => None,
-    }
+    #[error("Expected type id for {expected_type_name} ({expected_type_id:02x}) when reading OptionalCFrame; got {actual_type_id:02x}")]
+    BadOptionalCFrameFormat {
+        expected_type_name: String,
+        expected_type_id: u8,
+        actual_type_id: u8,
+    },
 }
 
 pub(crate) fn decode<R: Read>(reader: R) -> Result<WeakDom, Error> {
@@ -278,6 +159,11 @@ struct BinaryDeserializer<R> {
     /// Referents for all of the instances with no parent, in order they appear
     /// in the file.
     root_instance_refs: Vec<i32>,
+
+    /// Contains a set of unknown type IDs that we've encountered so far while
+    /// deserializing this file. We use this map in order to ensure we only
+    /// print one warning per unknown type ID when deserializing a file.
+    unknown_type_ids: HashSet<u8>,
 }
 
 /// All the information contained in the header before any chunks are read from
@@ -334,6 +220,7 @@ impl<R: Read> BinaryDeserializer<R> {
             type_infos,
             instances_by_ref,
             root_instance_refs: Vec::new(),
+            unknown_type_ids: HashSet::new(),
         })
     }
 
@@ -416,12 +303,40 @@ impl<R: Read> BinaryDeserializer<R> {
     fn decode_prop_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
         let type_id = chunk.read_le_u32()?;
         let prop_name = chunk.read_string()?;
-        let binary_type: Type = chunk.read_u8()?.try_into()?;
 
         let type_info = self
             .type_infos
             .get(&type_id)
             .ok_or(InnerError::InvalidTypeId { type_id })?;
+
+        // PROP chunks that contain no type byte are ignored by Roblox. This can
+        // happen when a new type is introduced.
+        //
+        // On 2021-04-08, OptionalCoordinateFrame was introduced, but its
+        // serialized format was just a type ID followed by the prop name. This
+        // leads us to believe that Roblox will silently ignore any PROP chunks
+        // that end immediately after the prop name, so we do the same.
+        let binary_type_byte = match chunk.read_u8() {
+            Ok(byte) => byte,
+            Err(_) => return Ok(()),
+        };
+
+        let binary_type: Type = match binary_type_byte.try_into() {
+            Ok(ty) => ty,
+            Err(_) => {
+                if self.unknown_type_ids.insert(binary_type_byte) {
+                    log::warn!(
+                        "Unknown value type ID {byte:#04x} ({byte}) in Roblox \
+                         binary model file. Found in property {class}.{prop}.",
+                        byte = binary_type_byte,
+                        class = type_info.type_name,
+                        prop = prop_name,
+                    );
+                }
+
+                return Ok(());
+            }
+        };
 
         log::trace!(
             "PROP chunk ({}.{}, instance type {}, prop type {}",
@@ -874,16 +789,14 @@ impl<R: Read> BinaryDeserializer<R> {
                                     chunk.read_le_f32()?,
                                 ),
                             ));
+                        } else if let Some(basic_rotation) = cframe::from_basic_rotation_id(id) {
+                            rotations.push(basic_rotation);
                         } else {
-                            let special_case = special_case_to_rotation(id).ok_or_else(|| {
-                                InnerError::BadCFrameOrientationId {
-                                    type_name: type_info.type_name.clone(),
-                                    prop_name: prop_name.clone(),
-                                    id,
-                                }
-                            })?;
-
-                            rotations.push(special_case);
+                            return Err(InnerError::BadRotationId {
+                                type_name: type_info.type_name.clone(),
+                                prop_name,
+                                id,
+                            });
                         }
                     }
 
@@ -1212,6 +1125,102 @@ impl<R: Read> BinaryDeserializer<R> {
                         valid_type_names: "SharedString",
                         actual_type_name: format!("{:?}", invalid_type),
                     })
+                }
+            },
+            Type::OptionalCFrame => match canonical_type {
+                VariantType::OptionalCFrame => {
+                    let referents = &type_info.referents;
+                    let mut rotations = Vec::with_capacity(referents.len());
+
+                    // Roblox writes a type marker for CFrame here that we don't
+                    // need to use. We explicitly check for this right now just
+                    // in case we're wrong and we do need it!
+                    let actual_type_id = chunk.read_u8()?;
+                    if actual_type_id != Type::CFrame as u8 {
+                        return Err(InnerError::BadOptionalCFrameFormat {
+                            expected_type_name: String::from("CFrame"),
+                            expected_type_id: Type::CFrame as u8,
+                            actual_type_id,
+                        });
+                    }
+
+                    for _ in 0..referents.len() {
+                        let id = chunk.read_u8()?;
+                        if id == 0 {
+                            rotations.push(Matrix3::new(
+                                Vector3::new(
+                                    chunk.read_le_f32()?,
+                                    chunk.read_le_f32()?,
+                                    chunk.read_le_f32()?,
+                                ),
+                                Vector3::new(
+                                    chunk.read_le_f32()?,
+                                    chunk.read_le_f32()?,
+                                    chunk.read_le_f32()?,
+                                ),
+                                Vector3::new(
+                                    chunk.read_le_f32()?,
+                                    chunk.read_le_f32()?,
+                                    chunk.read_le_f32()?,
+                                ),
+                            ));
+                        } else if let Some(basic_rotation) = cframe::from_basic_rotation_id(id) {
+                            rotations.push(basic_rotation);
+                        } else {
+                            return Err(InnerError::BadRotationId {
+                                type_name: type_info.type_name.clone(),
+                                prop_name,
+                                id,
+                            });
+                        }
+                    }
+
+                    let mut x = vec![0.0; referents.len()];
+                    let mut y = vec![0.0; referents.len()];
+                    let mut z = vec![0.0; referents.len()];
+
+                    chunk.read_interleaved_f32_array(&mut x)?;
+                    chunk.read_interleaved_f32_array(&mut y)?;
+                    chunk.read_interleaved_f32_array(&mut z)?;
+
+                    // Roblox writes a type marker for Bool here that we don't
+                    // need to use. We explicitly check for this right now just
+                    // in case we're wrong and we do need it!
+                    let actual_type_id = chunk.read_u8()?;
+                    if actual_type_id != Type::Bool as u8 {
+                        return Err(InnerError::BadOptionalCFrameFormat {
+                            expected_type_name: String::from("Bool"),
+                            expected_type_id: Type::Bool as u8,
+                            actual_type_id,
+                        });
+                    }
+
+                    let values = x
+                        .into_iter()
+                        .zip(y)
+                        .zip(z)
+                        .map(|((x, y), z)| Vector3::new(x, y, z))
+                        .zip(rotations)
+                        .map(|(position, rotation)| {
+                            if chunk.read_u8().ok()? == 0 {
+                                None
+                            } else {
+                                Some(CFrame::new(position, rotation))
+                            }
+                        });
+
+                    for (cframe, referent) in values.zip(referents) {
+                        let instance = self.instances_by_ref.get_mut(referent).unwrap();
+                        instance.builder.add_property(&canonical_name, cframe);
+                    }
+                }
+                invalid_type => {
+                    return Err(InnerError::PropTypeMismatch {
+                        type_name: type_info.type_name.clone(),
+                        prop_name,
+                        valid_type_names: "OptionalCFrame",
+                        actual_type_name: format!("{:?}", invalid_type),
+                    });
                 }
             },
         }