+++ rbx_types/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased Changes
 * Added an `Attributes` struct to facilitate reading and writing of attribute values.
 
+## 0.3.1 (2021-05-14)
+* Added `Variant::OptionalCoordinateFrame`.
+
 ## 0.3.0 (2021-02-16)
 * Renamed `EnumValue` to `Enum`.
 * Added `Display` and `FromStr` implementations for `Ref`.